services/strategy_runtime/requirements.txt
confluent-kafka
psycopg2-binary
python-dotenv
pandas
numpy
fastapi
uvicorn
requests